[refact]파일구조 분리 및 함수분리
